+++ Classes/DataHandling/DataHandler.php
@@ -33,6 +33,7 @@ use TYPO3\CMS\Backend\Utility\BackendUtility;
 use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\DataHandling\Localization\DataMapProcessor;
+use TYPO3\CMS\Core\DataHandling\SlugEnricher;
 use TYPO3\CMS\Core\Utility\ArrayUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\MathUtility;
@@ -56,7 +57,7 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
      * Processing the data-array
      * Call this function to process the data-array set by start()
      *
-     * @return void|false
+     * @return bool|void
      */
     public function process_datamap()
     {
@@ -68,23 +69,20 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
         $this->datamap = $this->unsetElementsToBeDeleted($this->datamap);
         // Editing frozen:
         if ($this->BE_USER->workspace !== 0 && $this->BE_USER->workspaceRec['freeze']) {
-            if ($this->enableLogging) {
-                $this->newlog('All editing in this workspace has been frozen!', 1);
-            }
+            $this->newlog('All editing in this workspace has been frozen!', 1);
             return false;
         }
         // First prepare user defined objects (if any) for hooks which extend this function:
         $hookObjectsArr = [];
-        if (is_array($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass'])) {
-            foreach ($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass'] as $classRef) {
-                $hookObject = GeneralUtility::getUserObj($classRef);
-                if (method_exists($hookObject, 'processDatamap_beforeStart')) {
-                    $hookObject->processDatamap_beforeStart($this);
-                }
-                $hookObjectsArr[] = $hookObject;
+        foreach ($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass'] ?? [] as $className) {
+            $hookObject = GeneralUtility::makeInstance($className);
+            if (method_exists($hookObject, 'processDatamap_beforeStart')) {
+                $hookObject->processDatamap_beforeStart($this);
             }
+            $hookObjectsArr[] = $hookObject;
         }
         // Pre-process data-map and synchronize localization states
+        $this->datamap = GeneralUtility::makeInstance(SlugEnricher::class)->enrichDataMap($this->datamap);
         $this->datamap = DataMapProcessor::instance($this->datamap, $this->BE_USER)->process();
         // Organize tables so that the pages-table is always processed first. This is required if you want to make sure that content pointing to a new page will be created.
         $orderOfTables = [];
@@ -101,7 +99,7 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
             //	   - the table is set with content in the data-array (if not, there's nothing to process...)
             //	   - permissions for tableaccess OK
             $modifyAccessList = $this->checkModifyAccessList($table);
-            if ($this->enableLogging && !$modifyAccessList) {
+            if (!$modifyAccessList) {
                 $this->log($table, 0, 2, 0, 1, 'Attempt to modify table \'%s\' without permission', 1, [$table]);
             }
             if (!isset($GLOBALS['TCA'][$table]) || $this->tableReadOnly($table) || !is_array($this->datamap[$table]) || !$modifyAccessList) {
@@ -139,12 +137,11 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                     $fieldArray = $this->newFieldArray($table);
                     // A pid must be set for new records.
                     if (isset($incomingFieldArray['pid'])) {
-                        // $value = the pid
                         $pid_value = $incomingFieldArray['pid'];
                         // Checking and finding numerical pid, it may be a string-reference to another value
-                        $OK = 1;
+                        $canProceed = true;
                         // If a NEW... id
-                        if (strstr($pid_value, 'NEW')) {
+                        if (strpos($pid_value, 'NEW') !== false) {
                             if ($pid_value[0] === '-') {
                                 $negFlag = -1;
                                 $pid_value = substr($pid_value, 1);
@@ -158,68 +155,36 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                                 }
                                 $pid_value = (int)($negFlag * $this->substNEWwithIDs[$pid_value]);
                             } else {
-                                $OK = 0;
+                                $canProceed = false;
                             }
                         }
                         $pid_value = (int)$pid_value;
-                        // The $pid_value is now the numerical pid at this point
-                        if ($OK) {
-                            $sortRow = $GLOBALS['TCA'][$table]['ctrl']['sortby'];
-                            // Points to a page on which to insert the element, possibly in the top of the page
-                            if ($pid_value >= 0) {
-                                // If this table is sorted we better find the top sorting number
-                                if ($sortRow) {
-                                    $fieldArray[$sortRow] = $this->getSortNumber($table, 0, $pid_value);
-                                }
-                                // The numerical pid is inserted in the data array
-                                $fieldArray['pid'] = $pid_value;
-                            } else {
-                                // points to another record before ifself
-                                // If this table is sorted we better find the top sorting number
-                                if ($sortRow) {
-                                    // Because $pid_value is < 0, getSortNumber returns an array
-                                    $tempArray = $this->getSortNumber($table, 0, $pid_value);
-                                    $fieldArray['pid'] = $tempArray['pid'];
-                                    $fieldArray[$sortRow] = $tempArray['sortNumber'];
-                                } else {
-                                    // Here we fetch the PID of the record that we point to...
-                                    $tempdata = $this->recordInfo($table, abs($pid_value), 'pid');
-                                    $fieldArray['pid'] = $tempdata['pid'];
-                                }
-                            }
+                        if ($canProceed) {
+                            $fieldArray = $this->resolveSortingAndPidForNewRecord($table, $pid_value, $fieldArray);
                         }
                     }
                     $theRealPid = $fieldArray['pid'];
                     // Now, check if we may insert records on this pid.
                     if ($theRealPid >= 0) {
                         // Checks if records can be inserted on this $pid.
-                        $recordAccess = $this->checkRecordInsertAccess($table, $theRealPid);
+                        // If this is a page translation, the check needs to be done for the l10n_parent record
+                        if ($table === 'pages' && $incomingFieldArray[$GLOBALS['TCA'][$table]['ctrl']['languageField']] > 0 && $incomingFieldArray[$GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField']] > 0) {
+                            $recordAccess = $this->checkRecordInsertAccess($table, $incomingFieldArray[$GLOBALS['TCA'][$table]['ctrl']['transOrigPointerField']]);
+                        } else {
+                            $recordAccess = $this->checkRecordInsertAccess($table, $theRealPid);
+                        }
                         if ($recordAccess) {
                             $this->addDefaultPermittedLanguageIfNotSet($table, $incomingFieldArray);
-                            $recordAccess = $this->BE_USER->recordEditAccessInternals(
-                                $table,
-                                $incomingFieldArray,
-                                true
-                            );
+                            $recordAccess = $this->BE_USER->recordEditAccessInternals($table, $incomingFieldArray, true);
                             if (!$recordAccess) {
-                                if ($this->enableLogging) {
-                                    $this->newlog(
-                                        'recordEditAccessInternals() check failed. [' . $this->BE_USER->errorMsg . ']',
-                                        1
-                                    );
-                                }
+                                $this->newlog('recordEditAccessInternals() check failed. [' . $this->BE_USER->errorMsg . ']', 1);
                             } elseif (!$this->bypassWorkspaceRestrictions) {
                                 // Workspace related processing:
                                 // If LIVE records cannot be created in the current PID due to workspace restrictions, prepare creation of placeholder-record
                                 if ($res = $this->BE_USER->workspaceAllowLiveRecordsInPID($theRealPid, $table)) {
                                     if ($res < 0) {
                                         $recordAccess = false;
-                                        if ($this->enableLogging) {
-                                            $this->newlog(
-                                                'Stage for versioning root point and users access level did not allow for editing',
-                                                1
-                                            );
-                                        }
+                                        $this->newlog('Stage for versioning root point and users access level did not allow for editing', 1);
                                     }
                                 } else {
                                     // So, if no live records were allowed, we have to create a new version of this record:
@@ -227,18 +192,13 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                                         $createNewVersion = true;
                                     } else {
                                         $recordAccess = false;
-                                        if ($this->enableLogging) {
-                                            $this->newlog(
-                                                'Record could not be created in this workspace in this branch',
-                                                1
-                                            );
-                                        }
+                                        $this->newlog('Record could not be created in this workspace in this branch', 1);
                                     }
                                 }
                             }
                         }
                     } else {
-                        debug('Internal ERROR: pid should not be less than zero!');
+                        $this->logger->debug('Internal ERROR: pid should not be less than zero!');
                     }
                     // Yes new record, change $record_status to 'insert'
                     $status = 'new';
@@ -249,29 +209,21 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                     if (!$recordAccess) {
                         if ($this->enableLogging) {
                             $propArr = $this->getRecordProperties($table, $id);
-                            $this->log($table, $id, 2, 0, 1,
-                                'Attempt to modify record \'%s\' (%s) without permission. Or non-existing page.', 2,
-                                [$propArr['header'], $table . ':' . $id], $propArr['event_pid']);
+                            $this->log($table, $id, 2, 0, 1, 'Attempt to modify record \'%s\' (%s) without permission. Or non-existing page.', 2, [$propArr['header'], $table . ':' . $id], $propArr['event_pid']);
                         }
                         continue;
                     }
                     // Next check of the record permissions (internals)
                     $recordAccess = $this->BE_USER->recordEditAccessInternals($table, $id);
                     if (!$recordAccess) {
-                        if ($this->enableLogging) {
-                            $this->newlog(
-                                'recordEditAccessInternals() check failed. [' . $this->BE_USER->errorMsg . ']',
-                                1
-                            );
-                        }
+                        $this->newlog('recordEditAccessInternals() check failed. [' . $this->BE_USER->errorMsg . ']', 1);
                     } else {
                         // Here we fetch the PID of the record that we point to...
-                        $tempdata = $this->recordInfo($table, $id,
-                            'pid' . ($GLOBALS['TCA'][$table]['ctrl']['versioningWS'] ? ',t3ver_wsid,t3ver_stage' : ''));
+                        $tempdata = $this->recordInfo($table, $id, 'pid' . (!empty($GLOBALS['TCA'][$table]['ctrl']['versioningWS']) ? ',t3ver_wsid,t3ver_stage' : ''));
                         $theRealPid = $tempdata['pid'];
                         // Use the new id of the versionized record we're trying to write to:
                         // (This record is a child record of a parent and has already been versionized.)
-                        if ($this->autoVersionIdMap[$table][$id]) {
+                        if (!empty($this->autoVersionIdMap[$table][$id])) {
                             // For the reason that creating a new version of this record, automatically
                             // created related child records (e.g. "IRRE"), update the accordant field:
                             $this->getVersionizedIncomingFieldArray($table, $id, $incomingFieldArray, $registerDBList);
@@ -279,13 +231,11 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                             $id = $this->autoVersionIdMap[$table][$id];
                             $recordAccess = true;
                             $this->autoVersioningUpdate = true;
-                        } elseif (!$this->bypassWorkspaceRestrictions && ($errorCode = $this->BE_USER->workspaceCannotEditRecord($table,
-                                $tempdata))) {
+                        } elseif (!$this->bypassWorkspaceRestrictions && ($errorCode = $this->BE_USER->workspaceCannotEditRecord($table, $tempdata))) {
                             $recordAccess = false;
                             // Versioning is required and it must be offline version!
                             // Check if there already is a workspace version
-                            $WSversion = BackendUtility::getWorkspaceVersionOfRecord($this->BE_USER->workspace, $table,
-                                $id, 'uid,t3ver_oid');
+                            $WSversion = BackendUtility::getWorkspaceVersionOfRecord($this->BE_USER->workspace, $table, $id, 'uid,t3ver_oid');
                             if ($WSversion) {
                                 $id = $WSversion['uid'];
                                 $recordAccess = true;
@@ -293,7 +243,7 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                                 // new version of a record created in a workspace - so always refresh pagetree to indicate there is a change in the workspace
                                 $this->pagetreeNeedsRefresh = true;
 
-                                /** @var $tce DataHandler */
+                                /** @var DataHandler $tce */
                                 $tce = GeneralUtility::makeInstance(__CLASS__);
                                 $tce->enableLogging = $this->enableLogging;
                                 // Setting up command for creating a new version of the record:
@@ -303,21 +253,18 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                                     // Default is to create a version of the individual records... element versioning that is.
                                     'label' => 'Auto-created for WS #' . $this->BE_USER->workspace
                                 ];
-                                $tce->start([], $cmd);
+                                $tce->start([], $cmd, $this->BE_USER);
                                 $tce->process_cmdmap();
                                 $this->errorLog = array_merge($this->errorLog, $tce->errorLog);
                                 // If copying was successful, share the new uids (also of related children):
-                                if ($tce->copyMappingArray[$table][$id]) {
+                                if (!empty($tce->copyMappingArray[$table][$id])) {
                                     foreach ($tce->copyMappingArray as $origTable => $origIdArray) {
                                         foreach ($origIdArray as $origId => $newId) {
                                             $this->uploadedFileArray[$origTable][$newId] = $this->uploadedFileArray[$origTable][$origId];
                                             $this->autoVersionIdMap[$origTable][$origId] = $newId;
                                         }
                                     }
-                                    ArrayUtility::mergeRecursiveWithOverrule(
-                                        $this->RTEmagic_copyIndex,
-                                        $tce->RTEmagic_copyIndex
-                                    );
+                                    ArrayUtility::mergeRecursiveWithOverrule($this->RTEmagic_copyIndex, $tce->RTEmagic_copyIndex);
                                     // See where RTEmagic_copyIndex is used inside fillInFieldArray() for more information...
                                     // Update registerDBList, that holds the copied relations to child records:
                                     $registerDBList = array_merge($registerDBList, $tce->registerDBList);
@@ -328,17 +275,11 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                                     $id = $this->autoVersionIdMap[$table][$id];
                                     $recordAccess = true;
                                     $this->autoVersioningUpdate = true;
-                                } elseif ($this->enableLogging) {
-                                    $this->newlog(
-                                        'Could not be edited in offline workspace in the branch where found (failure state: \'' . $errorCode . '\'). Auto-creation of version failed!',
-                                        1
-                                    );
+                                } else {
+                                    $this->newlog('Could not be edited in offline workspace in the branch where found (failure state: \'' . $errorCode . '\'). Auto-creation of version failed!', 1);
                                 }
-                            } elseif ($this->enableLogging) {
-                                $this->newlog(
-                                    'Could not be edited in offline workspace in the branch where found (failure state: \'' . $errorCode . '\'). Auto-creation of version not allowed in workspace!',
-                                    1
-                                );
+                            } else {
+                                $this->newlog('Could not be edited in offline workspace in the branch where found (failure state: \'' . $errorCode . '\'). Auto-creation of version not allowed in workspace!', 1);
                             }
                         }
                     }
@@ -351,17 +292,15 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                 }
 
                 // Here the "pid" is set IF NOT the old pid was a string pointing to a place in the subst-id array.
-                list($tscPID) = BackendUtility::getTSCpid($table, $id,
-                    $old_pid_value ? $old_pid_value : $fieldArray['pid']);
+                list($tscPID) = BackendUtility::getTSCpid($table, $id, $old_pid_value ? $old_pid_value : $fieldArray['pid']);
                 if ($status === 'new' && $table === 'pages') {
-                    $TSConfig = $this->getTCEMAIN_TSconfig($tscPID);
+                    $TSConfig = BackendUtility::getPagesTSconfig($tscPID)['TCEMAIN.'] ?? [];
                     if (isset($TSConfig['permissions.']) && is_array($TSConfig['permissions.'])) {
                         $fieldArray = $this->setTSconfigPermissions($fieldArray, $TSConfig['permissions.']);
                     }
                 }
                 // Processing of all fields in incomingFieldArray and setting them in $fieldArray
-                $fieldArray = $this->fillInFieldArray($table, $id, $fieldArray, $incomingFieldArray, $theRealPid,
-                    $status, $tscPID);
+                $fieldArray = $this->fillInFieldArray($table, $id, $fieldArray, $incomingFieldArray, $theRealPid, $status, $tscPID);
                 $newVersion_placeholderFieldArray = [];
                 if ($createNewVersion) {
                     // create a placeholder array with already processed field content
@@ -429,10 +368,7 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                             $newVersion_placeholderFieldArray['t3ver_wsid'] = $this->BE_USER->workspace;
                             $newVersion_placeholderFieldArray[$GLOBALS['TCA'][$table]['ctrl']['label']] = $this->getPlaceholderTitleForTableLabel($table);
                             // Saving placeholder as 'original'
-                            // PxDbSequencer: Call it with the additional $suggestedUid parameter
-                            $this->insertDB($table, $id, $newVersion_placeholderFieldArray, false,
-                                (int)$incomingFieldArray['uid']);
-
+                            $this->insertDB($table, $id, $newVersion_placeholderFieldArray, false, $incomingFieldArray['uid'] ?? 0);
                             // For the actual new offline version, set versioning values to point to placeholder:
                             $fieldArray['pid'] = -1;
                             $fieldArray['t3ver_oid'] = $this->substNEWwithIDs[$id];
@@ -445,21 +381,18 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                             // PxDbSequencer: Reset $incomingFieldArray['uid']
                             $incomingFieldArray['uid'] = 0;
                             // PxDbSequencer: Re-call PxDbsequencer DataHandlerHook to generate a sequenced uid for the placeholder record, if needed
-                            $pxDbSequencerHookObj = GeneralUtility::getUserObj($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass']['tx_pxdbsequencer']);
+                            $pxDbSequencerHookObj = GeneralUtility::makeInstance($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass']['tx_pxdbsequencer']);
                             if (method_exists($pxDbSequencerHookObj, 'processDatamap_preProcessFieldArray')) {
-                                $pxDbSequencerHookObj->processDatamap_preProcessFieldArray($incomingFieldArray, $table,
-                                    $id, $this);
+                                $pxDbSequencerHookObj->processDatamap_preProcessFieldArray($incomingFieldArray, $table, $id, $this);
                             }
 
                             // When inserted, $this->substNEWwithIDs[$id] will be changed to the uid of THIS version and so the interface will pick it up just nice!
                             // PxDbSequencer: Call it with the additional $suggestedUid parameter
-                            $phShadowId = $this->insertDB($table, $id, $fieldArray, true,
-                                (int)$incomingFieldArray['uid'], true);
+                            $phShadowId = $this->insertDB($table, $id, $fieldArray, true, (int)$incomingFieldArray['uid'], true);
 
                             if ($phShadowId) {
                                 // Processes fields of the placeholder record:
-                                $this->triggerRemapAction($table, $id, [$this, 'placeholderShadowing'],
-                                    [$table, $phShadowId]);
+                                $this->triggerRemapAction($table, $id, [$this, 'placeholderShadowing'], [$table, $phShadowId]);
                                 // Hold auto-versionized ids of placeholders:
                                 $this->autoVersionIdMap[$table][$this->substNEWwithIDs[$id]] = $phShadowId;
                             }
@@ -470,8 +403,7 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                         if ($table === 'pages') {
                             // only a certain number of fields needs to be checked for updates
                             // if $this->checkSimilar is TRUE, fields with unchanged values are already removed here
-                            $fieldsToCheck = array_intersect($this->pagetreeRefreshFieldsFromPages,
-                                array_keys($fieldArray));
+                            $fieldsToCheck = array_intersect($this->pagetreeRefreshFieldsFromPages, array_keys($fieldArray));
                             if (!empty($fieldsToCheck)) {
                                 $this->pagetreeNeedsRefresh = true;
                             }
@@ -515,14 +447,8 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
      * @param bool $dontSetNewIdIndex If TRUE, the ->substNEWwithIDs array is not updated. Only useful in very rare circumstances!
      * @return int|null Returns ID on success.
      */
-    public function insertDB(
-        $table,
-        $id,
-        $fieldArray,
-        $newVersion = false,
-        $suggestedUid = 0,
-        $dontSetNewIdIndex = false
-    ) {
+    public function insertDB($table, $id, $fieldArray, $newVersion = false, $suggestedUid = 0, $dontSetNewIdIndex = false)
+    {
         if (is_array($fieldArray) && is_array($GLOBALS['TCA'][$table]) && isset($fieldArray['pid'])) {
             // Do NOT insert the UID field, ever!
             unset($fieldArray['uid']);
@@ -550,7 +476,6 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                 ) {
                     $typeArray[$GLOBALS['TCA'][$table]['ctrl']['transOrigDiffSourceField']] = Connection::PARAM_LOB;
                 }
-                /** @var Connection $connection */
                 $connection = GeneralUtility::makeInstance(ConnectionPool::class)->getConnectionForTable($table);
                 $insertErrorMessage = '';
                 try {
@@ -586,20 +511,20 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                     }
                     // Update reference index:
                     $this->updateRefIndex($table, $id);
+
+                    // Store in history
+                    $this->getRecordHistoryStore()->addRecord($table, $id, $newRow);
+
                     if ($newVersion) {
                         if ($this->enableLogging) {
                             $propArr = $this->getRecordPropertiesFromRow($table, $newRow);
-                            $this->log($table, $id, 1, 0, 0,
-                                'New version created of table \'%s\', uid \'%s\'. UID of new version is \'%s\'', 10,
-                                [$table, $fieldArray['t3ver_oid'], $id], $propArr['event_pid'], $NEW_id);
+                            $this->log($table, $id, 1, 0, 0, 'New version created of table \'%s\', uid \'%s\'. UID of new version is \'%s\'', 10, [$table, $fieldArray['t3ver_oid'], $id], $propArr['event_pid'], $NEW_id);
                         }
                     } else {
                         if ($this->enableLogging) {
                             $propArr = $this->getRecordPropertiesFromRow($table, $newRow);
                             $page_propArr = $this->getRecordProperties('pages', $propArr['pid']);
-                            $this->log($table, $id, 1, 0, 0, 'Record \'%s\' (%s) was inserted on page \'%s\' (%s)', 10,
-                                [$propArr['header'], $table . ':' . $id, $page_propArr['header'], $newRow['pid']],
-                                $newRow['pid'], $NEW_id);
+                            $this->log($table, $id, 1, 0, 0, 'Record \'%s\' (%s) was inserted on page \'%s\' (%s)', 10, [$propArr['header'], $table . ':' . $id, $page_propArr['header'], $newRow['pid']], $newRow['pid'], $NEW_id);
                         }
                         // Clear cache for relevant pages:
                         $this->registerRecordIdForPageCacheClearing($table, $id);
@@ -607,8 +532,7 @@ class DataHandler extends \TYPO3\CMS\Core\DataHandling\DataHandler
                     return $id;
                 }
                 if ($this->enableLogging) {
-                    $this->log($table, $id, 1, 0, 2, 'SQL error: \'%s\' (%s)', 12,
-                        [$insertErrorMessage, $table . ':' . $id]);
+                    $this->log($table, $id, 1, 0, 2, 'SQL error: \'%s\' (%s)', 12, [$insertErrorMessage, $table . ':' . $id]);
                 }
             }
         }